+++ src/gui/GraphicsCursorItem.h
@@ -14,7 +14,7 @@ class GraphicsCursorItem : public QObject, public QGraphicsItem
     Q_OBJECT
 public:
     GraphicsCursorItem( int height, const QPen& pen );
-    int cursorPos() const { return pos().x(); }
+    int cursorPos() const { return ( int )pos().x(); }
     void setCursorPos( int position );
     virtual QRectF boundingRect() const;
